Добавил флаги в CMake, исправил недочёты

+++ src/adress.cpp
@@ -11,7 +11,7 @@ std::string Adress::GetStreet() {
 // Returnrs true if string are equal (ignores lowcase)
 bool Adress::CompareString(const std::string &lhs, const std::string &rhs) {
 	if (lhs.size() == rhs.size()) {
-		for (int i = 0; i < lhs.size(); ++i) {
+		for (size_t i = 0; i < lhs.size(); ++i) {
 			if (std::tolower(lhs[i]) != std::tolower(rhs[i])) {
 				return false;
 			}
@@ -23,8 +23,7 @@ bool Adress::CompareString(const std::string &lhs, const std::string &rhs) {
 }
 
 bool Adress::IsEqual(const Adress &lhs, const Adress &rhs) {
-	return CompareString(lhs.City, rhs.City) && CompareString(lhs.Street, rhs.Street)
-		&& (lhs.House == rhs.House) && (lhs.Flat == rhs.Flat);
+	return CompareString(lhs.City, rhs.City) && CompareString(lhs.Street, rhs.Street) && (lhs.House == rhs.House) && (lhs.Flat == rhs.Flat);
 }
 
 bool Adress::IsInCity(const Adress &adress, const std::string &city, const std::string &street) {
@@ -32,6 +31,5 @@ bool Adress::IsInCity(const Adress &adress, const std::string &city, const std::
 }
 
 bool Adress::Neighbors(const Adress &lhs, const Adress &rhs) {
-	return CompareString(lhs.City, rhs.City) && CompareString(lhs.Street, rhs.Street)
-		&& (std::abs((int)lhs.House - (int)rhs.House) == 2 or rhs.House == lhs.House);
+	return CompareString(lhs.City, rhs.City) && CompareString(lhs.Street, rhs.Street) && (std::abs((int)lhs.House - (int)rhs.House) == 2 or rhs.House == lhs.House);
 }
